webscraper_requirements.txt
requests
pillow
bs4
selenium
